DRTII-1382 Fix shift update message for prod app. Move test endpoint to test controller. Rename
